Moved ConnectionData related data and functions to separate translation unit for later re-use
